Updated isEmpty method in NumStack
The Stack object created in NumStack will use the size method to
determine if NumStack is empty or not.
The test i made will check if stack is empty on creation.
I also noticed that i didnt do the fake test for push, so i wrote the
code for it.