Added functionality to pop method and add javadoc.
Commit Description:
When pop method is called in numStack it will return the newest float
value.
I also put javadoc for the NumStack class.
Test:
I updated the faking pop test, by first pushing a value the seeing if it
pops the same value. I aslo checked to see if numStack is empty after
the pop in the same test.