Starting to Implement Reverse Polish Calculatior.
Commit Description:
I have created a TestRevPolishCalc and RevPolishClac class along with a
calculator interface, which the RevPolishCalc will implement later.

Testing:
No Testing for this commit.

Feature Branch Description:
The aim of this branch will be to use the NumStack class implemented in
the previous feature branch, as an attribute for the RevPolishCalc class to
try and evealuate string expressions in reverse polish.