Evaluate method can do addition on any two numbers
Commit Description:
Ive updated the evaluate method, so that now it can do addition on two
different numbers. This was done by splitting string expression to an
String array and hard coded the index of values i wanted from the array.

Testing:
I made a second test for addition of two numbers which were
different from the first test. This failed my test because i faked the
results earlier. Ive modified my code as described in the 'Commit Description' to then pass this test.