Evalaute does addition on multiple numbers
Commit Description\Testing:
I first write a new test where addition is being done on ten thousand
numbers in a loop.
This test fails.
I Then update the code in evaluate so that it can handle multiple
numbers when doing addition, using regular expressions,  an array and
the Math libary.
The test then passed.

Bug Fix:
In NumStack there was a mistake in the push method where an int value was sent
to the Entry class, but it was meant to be a float value instead.
This was a problem beacuse it caused rounding errors when adding
numbers.
I had fixed this bug by passing a float.

Extra/Refactoring:
Added some comments in evaluate method.