Evaluate method can now subtract numbers.
Commit Description/Testing:
I make a test where i subtract to numbers for the evaluate method.
The test fails.
I then update the evaluate method by pushing when ever it detects a
number in the string expression and popping twice when it detects an
operator.
I also made a new method called doOperation, so when evaluate detects an
operater in the string it will call this method to do the calculations
on the two popped values, then push the result back on the stack.
The test the passed.

Extra/Refactoring:
Got rid of newline on javadoc in BadTypeException class.