Evaluate can now identify negative numbers.
Commit Description/Testing:
I wrote a few test that have negative numbers in there string
expression.
The test failed, this is because the programme will look at the negative
sign and treat it as a operator.
To solve this problem i updated my regex string, so that it also covers
negative numbers to.
The test the passed.