Giving GUIView backend functionality
Commit Description:
I first create a new class called Observer, this class will be used for
the Contoller class i will develop later. The observer class just makes
that when GUIView sends an update of a new expression the Contoller eill
be notified.

In the GUIView class i first define some of the componets in the screen,
then reference the answer component so that that the setAnswer method
can update its value.
I also define the input component so that the getExpression method can
retrive the user input.
I also define a submit component, so that the addCalcObserver method can
notify the observer when user clicks submit.