View and Assembly now linked using an Controller
Commit Description:
I have now created the Controller class which has a filed for model
which represents the calculator type and view which represnts the type
of view used.
I first create an addView method which associates a view which a
contollers refrence to the calculate method.
The calaulate method calls the evaluate method from the model feild.

Extra:
I forgot to add the updated Javafx fxml file so i put it in this commit.