Got rid of radio buttons in the application
Usability Testing:
The infix and postfix button dosent do anything because the calculator
will always be in postfix mode as thats the only mode which is currently
implemented.
There is no reason for the client to see these buttons for this version
of the calculator, so i have deleted them.