Fix bug where errors were not displayed on gui.
Usabllity Testing:
During usability testing i noticed that when a type a invalid exception
nothing gets output to the gui.
The reason for this is because veiw.setAnswer expects an String as an
argument but will get an exception instead.
To solve this i cath the exception and output a error message in string
format.