Bug Fix when user enters numbers but no operands
Commit Description/Testing:
I performed more usability testing to check that when a user types an
expression such as "10 10" it should output an invalidException. However
the calculator does not do this.
So i the write a test case for this in TestNumStack which fails.
I fix the code in controller and NumStack to look out for this case.
The test then passed.