Add clear method for stack to fix a bug
Commit Desciption:
As i was doing usability testing i encouterd a bug where error was not
being shown to user after typing an invalid excception. The reason for
this is beacuse a perivous invalid expression still has its values in
the stack so when i type a new expression the output will be incorrect
as it uses previous values.
So to fix this i make a clear method.

Testing:
I first create a test case clear which tests to see if stack is cleared
after calling clear method.
Test fails because i havent implementented the method.
After i implement the method the test pass.

Extra/Refactoring:
I add javadoc for the clear method.
I add a constructor for the stack class, it does not need to be tested
as test cases for this have been made indirectly.