Fix pop and push methods
Commit Description/TDD:
I update the previous test case to check if isEmpty returns true after
popping the one Entry that was pushed in the Stack.
The test fails beacuse the pop method was faked, so i update this by
using pop method from stack class.
I run the test again but still fails as the correct symbol was not being
output after popping.
This is beacuse my push method was also faked, to fix this i push
symbols which were passed as arguments.
The test then passes.