Create Constructor for Standard Calc
Commit Description/TDD:
I write a test in a class i created called TestStandardClac to see if a
constructor for StandardCalc class can be made.
The test fails, to fix this i just create a StandardCalc class and adda
constructor to it.

Feature Branch Description:
In this branch the aim is to get evaluations for infix working in the
assembly so that it can be connected to the GUI afterwards.