Evaluate can do times on two numbers
Commit Description/TDD:
I write a test for multiplication of two numbers.
The test fails beacuse evaluate did not look out for multiply, to solve
this i add a if statement to the evaluate method to make it look out for
multiply operator.

Extra:
I add a few extra asserts in the previous test cases to look out for
some different situations with addition and subtraction.