Evaluate can do add and times on muliple numbers
Commit Description/TDD:
I write a new test case to test expressions with addition and
subtractions on multiple numbers.
The test fails as evaluate could only do it with two numbers.
To make test pass, I update the evaluate method to use a while loop so that it can keep adding and multiplying lots of numbers.