Created peek method in opStack
Commit Description:
To further develop the evaluate method i must add a new method in
opStack called peek. This is to deal with precedence of operators when
converting to postfix.

TDD:
I make a new test case in TestOpStack, to peek the top value of its
stack.
The test fails because it has not been implemented, i then make the
method for it. The test the passes.