fix bug in peek method for opStack
Commit Description/TDD:
I update the previous test case to check that after peeking the value on
the stack, it does not remove it.
The test fails as i wrote pop stack inside the peek method.
To fix this i chage it to use the top method i made for stack.