Evalaulate now works with brackets
Commit Description/TDD:
I make a test to see if evaulate works with expressions that contain
brackets.
The test fails because evaulate does not take this into account.
TO fix this, I update evaluate to look out for bracket by pushing the
open bracket. Then when it sees a close bracket operator it will keep
popping until it gets to the open bracket operator.