Add CalcModel state class
Commit Description:
Before connecting the standard clac assembly code to the GUI, i will
first add a state class called CalcModel, which will switch between
postfix and infix. To avoid the need of if statements.

TDD:
I first write a test case to initiate a class called calc model.
The test fails and to fix this i just, create a class called ClacModel.

Fearure Branch Description:
This feature will combine infix and postfix to the GUI, so that the user
will be able to select what type of expression they would like to
evaluate.