GUI now works with infix and postfix
Commit Description:
I update the controller so that i can add it as a type observer to also
obsserve the type, the calculator is in.
i.e it will see if its in postfix mode or infix code

I also refactored code for TestCalcModel and CalcModel to prevent
redundent code.