Postfix expressions now invlaid for Infix mode
Usability Testing:
There was an bug with infix mode where postfix expressions were valid
accepted as valid inputs. To fix this a boolean variable is set to
lookout for the order of the expressiosn.
i.e if there are two numbrs next to each other without an operator then throw an error as
this isnt possible in infix expressions.