Fix Postfix not working on ASCII
Commit Description:
There was a problem where when writing postfix expressions on the ASCII
mode, it will keep throwing an error. This is because every expression
written had a space in front of it. I took this empty space to account
for infix mode thats why it worked but not in postfix mode.
To fix this i trim the begging of the string.

Extra:
Add a assert case to look out for expressions with empty space at
beggining.