ADD case if user enters invalid command
Commit Description:
If the user enters an invalid command, they get the option to choose if
they want open up the help menu or not. I give them this as a option
because they might have typed an invalid commad by accident.
I also update the menu screen to make it more clearer on how to use the
application.