Release Notes for Release-3.0
Commit Description:
Update release notes for 3.0 talking about newest ASCII addition.

Complience Testing:
Complience testing has passed as when running the application there are no checkstyle warnings.

Usability Testing:
Tested if both postfix and infix calculations work properly on the
terminal, which it does.
Tested if invalid outputs are given when expected which it does.
Tested every command the user type on terminal, these work properly.
Wrote clear instructions in the menu screen to make it easier for the
user to use so that it makes user experience better.

Documentation Testing:
Javadoc and comments have been added for the every class in the
programme, following checkstyle rules.
CHNAGELOG and REALESENOTES are also up to date.

Configuration Tested:
This version of the application has been tested on windows and
linux(noMachine).

Branch Description:
I decide to make a release here as there has been an significant change
to the application where the ASCII View has been implemented, it works
just like the GUI but on terminal.