[Bug Fix]Use BigDecimal Instead of float
When i start adding really large numbers or divide really small numbers,
the prescion of the answer becomes incorrect.
This is because float is capped to a max and min value.
To fix this i change from type float to BigDecimal instead, as
BigDecimal is capable of dealing with really large numbers and small
numbers.