[FIX]Give ASCII an inital type
Usability Testing:
There was a problem with ASCII where if you dont set a inital type and
try to evalaute a expression, it will give a error. This is beacuse it
is not given an inital state, to fix this i edit the controller class
and set the inital state to postfix.