[Bug Fix] Fix Output Overflow
System Testing:
Fix bug where when ouptut has a number which is too large, it starts
overflowing and going over other components of the application.
To fix this i chnage the component of holding the anwer into a text
feild and not a text box, where the text feild box is disabled.