[Refactor]Move formatting away from controller
Remove code smell:
After evaluating the answer the controller will use decimal formater to
get rid of scientifc notation, then output the answer.
However its not the controllers job to do this, this jib should have
been done by the evaluate method in revPolishCalc.
So alter the code to make evaluate do this.

Testing:
I update the TDD test as i change the return type of the evaluate method
to string due to the decimal formatter. It shouldnt be a problem to make
the result a string as we dont need to perform anymore calculations on
the result anyway.